Enable running tests in parallel. Drop database when finished.
When the test database is first accessed by the unit tests,
a unique name is generated for the database. This name is
unique throughout the session and makes it possible to run
several test runs in parallel from different sessions.

The database is now also dropped when the test run finishes.